Migrate concat-tests to joint test file with `append()`
